ajust in SalesRecord

+++ TemDeTudo/Models/SalesRecord.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public Decimal Price { get; set; }
+        public double Price { get; set; }
         public SaleStatus Status { get; set; }
         public int SellerId { get; set; }
         public Seller Seller { get; set; }
@@ -12,7 +12,7 @@
     }
 
     public enum SaleStatus : int
-    {
+    { 
         PENDING = 0,
         BILLED = 1,
         CANCELED = 2
